backend/app/requirements.txt
alembic
python-dotenv
fastapi
sqlalchemy
uvicorn
pydantic
python-jose